Major update/rewrite...
Instead of using various userspace tools for replaygain analysis
and integrity checking etc, we now have a C extension called
aunalyzer (from audio analyzer) that gathers format information
of audio files and verifies their integrity through ffmpeg,
and also performs loudness analysis as per EBUR128 using
libebur128.

The loudness analysis results per track are used to calculate
replaygain tags following the replaygain2 spec (we get the same
track gain/peak results as with rsgain, that also uses ffmpeg +
libebur128). Together with the relative gating threshold used
during processing and exported to the python side and the accurate
track duration, we also attempt to get albun gain (which is not
as accurate as rsgain that integrates over all the states of the
tracks, we sum the sums in our case which is equivalent but we
loose some info info the process, in any case we are talking about
fractions of a dB in most cases, we add 0.5dB to be on the safe
side).

There is now support for resolving duplicates from the same album
(same album id, in the same folder), that may happen for example
if we end up commiting the same album multiple times in the
library, with e.g. different formats, or different pressings etc.
To do this we calculate a quality metric based on the file's
encoding (e.g. lossy vs lossless), sample rate, bit depth and a
dynamic range factor based on Loundess Range as per EBUR128. This
metric is used to sort duplicates, and the one with the best value
wins.

There were various cleanups here and there, and this is mostly
packaged now so it can be installed with pip. More work on that
will follow.

Changes were tested against a library snapshot and libguard worked
as expected, althoug more testing is required.

That's all for now...